All non-ENOENT errors should be handled the same
There's a lot of special-casing for EISDIR, EEXIST, etc.  It dawned on me
that we've special cased all the errors that fs.mkdir() can possibly raise,
so why not just make that the default: case, and cut out a bunch of code
that has to be read?

Argument against: if some error is ever raised, which *isn't* an error that
fs.mkdir() can 'possibly' raise, then it's raising an error that is *impossible*
and we probably want to get out asap.

But even in the case of impossible errors, if the dir is there, and it's
a dir, then who cares?  mkdir(1) ignores any errors making a directory with
-p if the directory exists and is a directory.